+++ .gitignore
@@ -1,6 +1,7 @@
 /.bundle/
 /.yardoc
 /Gemfile.lock
+/gemfiles/Gemfile-*.lock
 /_yardoc/
 /coverage/
 /doc/
