+++ cloudfront-rails.gemspec
@@ -20,12 +20,10 @@ Gem::Specification.new do |spec|
   spec.add_development_dependency "bundler", "~> 1.12"
   spec.add_development_dependency "rake", "~> 10.0"
   spec.add_development_dependency "rspec", "~> 3.0"
-  spec.add_development_dependency "rspec-rails"
 
   spec.add_development_dependency "webmock"
 
-  spec.add_dependency "rails", "~> 5.0"
-  spec.add_dependency "httparty", ">= 0.13.7"
+  spec.add_dependency "railties", "> 4.0"
 
   spec.required_ruby_version = ">= 2.0"
 end
