+++ gemfiles/Gemfile-5.x
@@ -1,0 +1,5 @@
+source 'https://rubygems.org'
+
+gemspec path: '..'
+
+gem 'railties', '~> 5.0'
