+++ lib/cloudfront/rails.rb
@@ -1,3 +1,6 @@
+require "rails"
+require "net/http"
+
 require "cloudfront/rails/version"
 
 module Cloudfront
@@ -6,4 +9,6 @@ module Cloudfront
   end
 end
 
+require "cloudfront/rails/importer"
+require "cloudfront/rails/importer/response_error"
 require "cloudfront/rails/railtie"
